fix: attempt to correct fullscreen woes
This commit aims to replace the Godot-side full screen handling
with a browser-side fullscreen handling.

This should partially or completely address:

- #272
- #226
- #208

## What Was The Challenge?

Two main problems:

1. There were two competing systems to set the app fullscreen or
   exit fullscreen. Godot can set the canvas context to
   fullscreen, the browser can set the window to fullscreen, and
   both can un-fullscreen the canvas context or the window.
   Neither system is aware of the other.
2. When setting the canvas to full screen, Godot's UI
   interactable area does not correspond to the visual UI on
   screen. This happens even if the canvas is set as full screen
   through Javascript. I do not know if the failure is on Godot's
   end, the browser's end, or due to our CSS (it'd still be a
   bug, but would explain why it isn't more widespread).

It seems that when setting the _document_ as full screen (as
opposed to _the canvas context_), there's no UI problem.

Note that we cannot call the browser API to set fullscreen from
Godot (for security reasons, browsers disable the functionality),
so the button to set the document as full screen has to be an
html button

This commit addresses problem 1 (#226) by:
- making sure the browser polls fullscreen changes and dispatches
  a signal to Godot
- making sure Godot can listen to the signal

It also addresses #272 by:
- removing the fullscreen button in the Godot app

Finally, it addresses problem 2 (#208) by:
- removing the Godot fullscreen button in browser contexts
- removing the fullscreen Godot shortcut in browser contexts
- adding an overlay html button in the top right corner