fix: stopped tween in lesson5(DrawingTurtle in ExampleDrawAnySquare.tscn) caused issues when the tween was run again
To Replicate the Bug: Navigate in Lesson 5 to the example in which the
turtle draws a square of variable(set by a "Size" slider) length. Click
the button labelled "draw_square()". While the animation is running,
change the "Size" slider to a new number and click the button labelled "Reset". Click the
"draw_square()" button again and let the animation complete. You should
be able to spot the bug in the form of the turtle making unexpected and
unwanted movements.