fix: work around random issue where regexes are not properly initialized in Godot 3.4.4
Ive been experiencing errors that cannot be reproduced reliably that the regexes created in _init() did not exist. It happened 3 times over dozens of runs. Repeating the same steps that initially triggered the bug did not cause it to happen again once, each time it was in a different context.